pyproject.toml
[project]
name = "self-learning-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "coverage (>=7.10.6,<8.0.0)",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)"
]

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119
